tidy up ukernel headers
