+++ runtime/src/iree/builtins/ukernel/pack_tile.c
@@ -4,7 +4,7 @@
 // See https://llvm.org/LICENSE.txt for license information.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-#include "iree/builtins/ukernel/pack_tile.h"
+#include "iree/builtins/ukernel/pack_internal.h"
 
 #if defined(IREE_UK_ARCH_ARM_64)
 #include "iree/builtins/ukernel/arch/arm_64/pack_arm_64.h"
