+++ runtime/src/iree/builtins/ukernel/tools/BUILD.bazel
@@ -21,7 +21,7 @@ iree_runtime_cc_library(
         "//runtime/src/iree/base",
         "//runtime/src/iree/base/internal:cpu",
         "//runtime/src/iree/base/internal:synchronization",
-        "//runtime/src/iree/builtins/ukernel:headers",
+        "//runtime/src/iree/builtins/ukernel",
         "//runtime/src/iree/schemas:cpu_data",
     ],
 )
@@ -33,7 +33,7 @@ iree_runtime_cc_library(
     deps = [
         ":util",
         "//runtime/src/iree/base",
-        "//runtime/src/iree/builtins/ukernel:headers",
+        "//runtime/src/iree/builtins/ukernel",
         "//runtime/src/iree/schemas:cpu_data",
     ],
 )
@@ -45,7 +45,7 @@ iree_runtime_cc_library(
     deps = [
         ":util",
         "//runtime/src/iree/base",
-        "//runtime/src/iree/builtins/ukernel:headers",
+        "//runtime/src/iree/builtins/ukernel",
         "//runtime/src/iree/schemas:cpu_data",
         "//runtime/src/iree/testing:benchmark",
     ],
@@ -58,7 +58,7 @@ iree_runtime_cc_library(
     deps = [
         ":benchmark",
         "//runtime/src/iree/base",
-        "//runtime/src/iree/builtins/ukernel:headers",
+        "//runtime/src/iree/builtins/ukernel",
         "//runtime/src/iree/testing:benchmark",
     ],
 )
@@ -72,6 +72,7 @@ cc_binary_benchmark(
         "//runtime/src/iree/base",
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/builtins/ukernel",
+        "//runtime/src/iree/builtins/ukernel:internal_headers",
         "//runtime/src/iree/testing:benchmark",
     ],
 )
@@ -85,6 +86,7 @@ iree_runtime_cc_test(
         "//runtime/src/iree/base",
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/builtins/ukernel",
+        "//runtime/src/iree/builtins/ukernel:internal_headers",
     ],
 )
 
@@ -98,6 +100,7 @@ cc_binary_benchmark(
         "//runtime/src/iree/base",
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/builtins/ukernel",
+        "//runtime/src/iree/builtins/ukernel:internal_headers",
         "//runtime/src/iree/testing:benchmark",
     ],
 )
@@ -111,6 +114,7 @@ iree_runtime_cc_test(
         "//runtime/src/iree/base",
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/builtins/ukernel",
+        "//runtime/src/iree/builtins/ukernel:internal_headers",
     ],
 )
 
@@ -124,6 +128,7 @@ cc_binary_benchmark(
         "//runtime/src/iree/base",
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/builtins/ukernel",
+        "//runtime/src/iree/builtins/ukernel:internal_headers",
         "//runtime/src/iree/testing:benchmark",
     ],
 )
@@ -138,6 +143,7 @@ iree_runtime_cc_test(
         "//runtime/src/iree/base/internal:cpu",
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/builtins/ukernel",
+        "//runtime/src/iree/builtins/ukernel:internal_headers",
     ],
 )
 
@@ -150,6 +156,7 @@ cc_binary_benchmark(
         "//runtime/src/iree/base",
         "//runtime/src/iree/base/internal:flags",
         "//runtime/src/iree/builtins/ukernel",
+        "//runtime/src/iree/builtins/ukernel:internal_headers",
         "//runtime/src/iree/testing:benchmark",
     ],
 )
