+++ runtime/src/iree/builtins/ukernel/tools/CMakeLists.txt
@@ -21,7 +21,7 @@ iree_cc_library(
     iree::base
     iree::base::internal::cpu
     iree::base::internal::synchronization
-    iree::builtins::ukernel::headers
+    iree::builtins::ukernel
     iree::schemas::cpu_data
   PUBLIC
 )
@@ -36,7 +36,7 @@ iree_cc_library(
   DEPS
     ::util
     iree::base
-    iree::builtins::ukernel::headers
+    iree::builtins::ukernel
     iree::schemas::cpu_data
   PUBLIC
 )
@@ -51,7 +51,7 @@ iree_cc_library(
   DEPS
     ::util
     iree::base
-    iree::builtins::ukernel::headers
+    iree::builtins::ukernel
     iree::schemas::cpu_data
     iree::testing::benchmark
   PUBLIC
@@ -67,7 +67,7 @@ iree_cc_library(
   DEPS
     ::benchmark
     iree::base
-    iree::builtins::ukernel::headers
+    iree::builtins::ukernel
     iree::testing::benchmark
   PUBLIC
 )
@@ -83,6 +83,7 @@ iree_cc_binary_benchmark(
     iree::base
     iree::base::internal::flags
     iree::builtins::ukernel
+    iree::builtins::ukernel::internal_headers
     iree::testing::benchmark
   TESTONLY
 )
@@ -98,6 +99,7 @@ iree_cc_test(
     iree::base
     iree::base::internal::flags
     iree::builtins::ukernel
+    iree::builtins::ukernel::internal_headers
 )
 
 iree_cc_binary_benchmark(
@@ -112,6 +114,7 @@ iree_cc_binary_benchmark(
     iree::base
     iree::base::internal::flags
     iree::builtins::ukernel
+    iree::builtins::ukernel::internal_headers
     iree::testing::benchmark
   TESTONLY
 )
@@ -127,6 +130,7 @@ iree_cc_test(
     iree::base
     iree::base::internal::flags
     iree::builtins::ukernel
+    iree::builtins::ukernel::internal_headers
 )
 
 iree_cc_binary_benchmark(
@@ -141,6 +145,7 @@ iree_cc_binary_benchmark(
     iree::base
     iree::base::internal::flags
     iree::builtins::ukernel
+    iree::builtins::ukernel::internal_headers
     iree::testing::benchmark
   TESTONLY
 )
@@ -157,6 +162,7 @@ iree_cc_test(
     iree::base::internal::cpu
     iree::base::internal::flags
     iree::builtins::ukernel
+    iree::builtins::ukernel::internal_headers
 )
 
 iree_cc_binary_benchmark(
@@ -170,6 +176,7 @@ iree_cc_binary_benchmark(
     iree::base
     iree::base::internal::flags
     iree::builtins::ukernel
+    iree::builtins::ukernel::internal_headers
     iree::testing::benchmark
   TESTONLY
 )
